\name{cugenV}
\alias{cugenV}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Uniform 0-1 Generator Visualization
}
\description{
Improved version of cugen, Can generate n random numbers.
}
\usage{
cugenV(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
n as number of random numbers in [0,1] interval.
}
}
\details{
-
}
\value{
This fanction returns a vector containing the random numbers.
}
\references{
www.wikipedia.com
}
\author{
P.Hessabi}
\note{
This function also plots the random numbers in order to visualize this distribution in [0,1] interval.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
cugenV.R
}
\examples{
cugenV(20000)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~cugenV }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~Nuniform }% __ONLY ONE__ keyword per line
